requirements.txt
python-gitlab
pygithub
